+++ EarTrumpet/DataModel/AppInformation/IAppInfo.cs
@@ -5,6 +5,7 @@ namespace EarTrumpet.DataModel.AppInformation
     public interface IAppInfo
     {
         event Action<IAppInfo> Stopped;
+        string AppId { get; }
         string DisplayName { get; }
         string ExeName { get; }
         string PackageInstallPath { get; }